Mark clearDevice test that requires disabled flag with a @DisabledFlag tag.
This test requires the FLAG_CLEAR_DEVICE_RESTRICTIONS disabled. In development, the flag is disabled by default. But in staging, the flag is enabled by default. This makes this test to fail as there is no exclusive marker requiring this test to either only run when the flag is disabled or disable the flag before running.

Bug: 417178868
Test: atest DeviceLockControllerRoboTests
Test: atest CtsDeviceLockTestCases
Flag: EXEMPT bugfix
Change-Id: I6e7516862a92ce9d0f46040fd77befb4eaddf37d