[DLC RECOL] Change FinalizationState from FINALIZED to UNFINALIZED when the device is provisioned.
Recollateralization introduces a new scenario where a device that had previously been finalized is provisioned again. To avoid an inconsistent state where the finalization state is FINALIZED for a device that is undergoing provisioning, we should should ensure that the finalization state is changed to UNFINALIZED when a device receives a PROVISION_READY status from the server.

Bug: 431915245
Test: atest DeviceLockControllerRoboTests
Flag: EXEMPT Behind the FeatureFlagProvider which defaults to false
Change-Id: Iac23df14709cab6c8bef7a6c940bf58f1d05419a