[DLC RECOL] Return device to clean state upon finalization.
When a device is finalized, the various stored parameters should be
cleared so that any subsequent provisioning attempt begins from a clean
state.

Also, we should be able to revert a device that was in the process
of provisioning but encountered a provisioning failure back to the clean
state it was in when it was finalized. The usage of these new methods in
ProvisionStateControllerImpl will be in subsequent CLs.

Bug: 433406463
Test: atest DeviceLockControllerRoboTests
Flag: EXEMPT Guarded by FeatureFlagProvider
Change-Id: I860c7cba0d3ecd0009b72505a601874c5c2b38d2