[DLC RECOL] Show recol-specific provision failure screens.
The ReportDeviceprovisionStateWorker might receive a BOTTOM_VIEW client
provision state from the server during a recollateralization
provisioning failure. If it does, the ProvisioningProgressViewModel will
be aware as it is observing the worker and the UI will be updated to
provide the user with a retry and an exit button.

Otherwise the
provision failure screen will not have these buttons and the client will
either automatically schedule a retry or check in again with the server
depending on if it received a RETRY, or CHECKIN response.

Bug: 433406463
Test: atest DeviceLockControllerRoboTests
Flag: EXEMPT Behind FeatureFlagProvider
Change-Id: I466822d11744f3ede7dcc38f5e8db172ab6b8e15