Fix[CheckIn]: Ensure phone package is enabled for check-in
Introduced a pre-condition for the `DeviceCheckInWorker` to ensure that the `com.android.phone` package is enabled before retrieving device ids.

Check-In depends on TelephonyService provided by the phone system app package. The package will be re-enabled or re-installed if the user had disabled or uninstalled it.In either case, user controls for the phone package are temporarily disabled then restored after device identifiers have been successfully retrieved. This pre-condition is turned OFF in aosp by default.

Bug: 433296069
Flag: EXEMPT fixing a vulnerability
Test: atest DeviceLockControllerRoboTests
Change-Id: I7124315fa8e773efdd598847337c3ab79d3951f1