[DLC RECOL] Reset mClearingInProgress variable
This variable is used to guard against race conditions (b/379925834)
however it remains set to true even after clearing has completed. This
causes issues if the device is subsequently re-enrolled and
re-provisioned - locking and unlocking will not work.

In this CL we ensure that the variable is reset to false after the
clearRestrictions operation completes.

FLAG: EXEMPT BUGFIX
Bug: 448579593
Test: atest DeviceLockControllerRoboTests

Change-Id: I990246ac205475cb768110c0d08084942859718e