[DLC RECOL] Fix provisioning failure exit and retry.
This CL fixes two bugs relating to provisioning failure in
recollateralization provisioning.

1) When a provisioning failure occurs and the user is presented with an
   exit and retry button and the user clicks on exit, a "check-in" FCM
message from the back end should trigger another provisioning flow.
This was not working because the finalization controller would retain
some in-memory state that would then prevent a subsequent provisioning
to kick off. We fix this by enforcing finalization state to
UNPROVISIONED at the start of every provisioning.

2) When a provisioning failure occurs and the user is presented with an
   exit and retry button and the user clicks on retry, the device should
re-attempt provisioning. This was not happening because of observing a
work info from a previous reportdeviceprovisioningstatus. This has been
fixed by not only cancelling but also pruning work during finalization.

Bug: 451870979
Bug: 451872347
Test: atest DeviceLockRoboElectricTests
Test: atest DeviceLockUnitTests
Test: atest DeviceLockControllerGoogleInstrumentationTests
Flag: EXEMPT BUGFIX
Change-Id: I045ff377f92fb3f9ae074ba1de7b722c4b77901f