Use system app dialer in lock task mode.
This addresses a vulnerability related to adding the default dialer to the allowlisted apps in lock task mode.

This CL instead adds the system dialer to the allowlisted apps. During provisioning, it also re-enables the system
dialer if it had been disabled by the user.

Bug: 452197710
Test: atest DeviceLockControllerGoogleInstrumentationTests
Test: atest DeviceLockControllerRoboTests
Test: Verified manually that the vulnerability is not reproducible with
the change.
Flag: EXEMPT BUGFIX

Change-Id: I4311fb1fa38b7ca0808ad7335f01d702c8775d89